requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
python-dotenv==1.0.0
requests==2.31.0
marshmallow==3.20.1
pytest==7.4.3
pytest-cov==4.1.0
hypothesis==6.92.1
WTForms==3.1.1
email-validator==2.1.0.post1
gunicorn==21.2.0
